Restructure project and update Line Bot instructions
