+++ AGENTS.md
@@ -2,7 +2,7 @@
 
 ## 1. 概述 (Overview)
 
-本技術規格檔定義了一個以 C# 與 Quartz.NET 實作之排程提醒 Agent，用於在指定「上課日期」前一日 21:00 及當日 07:00，自動向 LINE 群組推送文字提醒訊息。方案採用免費開源元件，並依照 OpenAI Codex 撰寫風格，以利自動化代碼產生。
+本技術規格檔定義了一個以 C# 與 Quartz.NET 實作之排程提醒 LINE Bot，能在指定「上課日期」前一日 21:00 及當日 07:00，向已存在的 LINE 群組推送文字訊息。此 Bot 使用 LINE Messaging API 的 Push 模式，不依賴 LINE Store 機制。
 
 ## 2. 功能需求 (Functional Requirements)
 
@@ -20,7 +20,13 @@
 - **可維護性**：採用清晰模組分層與 DI/配置管理。
 - **可擴充性**：未來可加入多筆日期、Flex Message、持久化 JobStore。
 
-## 4. 系統架構 (Architecture)
+## 4. LINE Bot 安裝與群組設定
+
+1. 在 LINE Developers 平台建立 Messaging API Channel，取得 Channel Access Token。
+2. 將 Bot 加入需要推播的 LINE 群組。
+3. 透過 Webhook 事件或訊息日誌取得 `groupId`，填入設定檔。
+
+## 5. 系統架構 (Architecture)
 
 ```
 +--------------+      +----------------+      +--------------------------+
@@ -39,16 +45,16 @@
 - **NotifyJob**：實作 IJob，執行時呼用 LINE Client 發送訊息。
 - **LINE Client**：透過 `HttpClient` 呼叫 `https://api.line.me/v2/bot/message/push`。
 
-## 5. 相依套件 (Dependencies)
+## 6. 相依套件 (Dependencies)
 
 - `Quartz` (Quartz.NET)
 - `Microsoft.Extensions.Hosting`
 - `Microsoft.Extensions.Configuration.Json`
 - `System.Net.Http`
 
-## 6. 設定檔 (Configuration)
+## 7. 設定檔 (Configuration)
 
-在專案根目錄新增 `appsettings.json`：
+在 `src/LineBotScheduler` 目錄新增 `appsettings.json`：
 
 ```json
 {
@@ -64,7 +70,7 @@
 - `ChannelAccessToken`：從 LINE Developers 取得。
 - `GroupId`：Webhook 事件 `source.groupId` 或手動取得。
 
-## 7. 模組說明 (Modules)
+## 8. 模組說明 (Modules)
 
 ### 7.1 Program.cs
 
@@ -87,7 +93,7 @@
   - 設定 Bearer Token、Header、Body。
   - 發送並處理回應。
 
-## 8. 排程流程 (Workflow)
+## 9. 排程流程 (Workflow)
 
 1. **使用者互動**：Console 輸入 `yyyy-MM-dd`。
 2. **日期驗證**：若格式錯誤，訊息提醒並結束。
@@ -107,7 +113,7 @@
 5. **Scheduler 啟動**：`await sched.Start();`
 6. **觸發執行**：到達時間，自動執行 `NotifyJob.Execute`
 
-## 9. 錯誤處理 (Error Handling)
+## 10. 錯誤處理 (Error Handling)
 
 - **日期解析失敗**：提示格式錯誤，退出程序。
 - **HTTP 呼叫失敗**：
@@ -115,7 +121,7 @@
   - 可選：加入重試策略（Polly）。
 - **Scheduler 啟動異常**：記錄並終止。
 
-## 10. 部署與執行 (Deployment & Execution)
+## 11. 部署與執行 (Deployment & Execution)
 
 - **開發測試**：在本機 Console 執行。
 - **生產環境**：
@@ -153,13 +159,13 @@
        ```
     4. **容器化**：撰寫 Dockerfile 並 `docker build/run`。
 
-## 11. 日誌與監控 (Logging & Monitoring)
+## 12. 日誌與監控 (Logging & Monitoring)
 
 - 建議整合 `Microsoft.Extensions.Logging`。
 - 輸出至檔案／控制台。
 - 可結合 ELK、Application Insights 等。
 
-## 12. 可擴充功能 (Extensibility)
+## 13. 可擴充功能 (Extensibility)
 
 - **多筆日期支援**：解析多行輸入，依序排程。
 - **Flex Message**：使用 JSON 模板推送更豐富的訊息樣板。
