+++ src/LineBotScheduler/NotifyJob.cs
@@ -4,7 +4,7 @@ using System.Text.Json;
 using Microsoft.Extensions.Options;
 using Quartz;
 
-namespace LineReminder
+namespace LineBotScheduler
 {
     public class NotifyJob : IJob
     {
